move `construct_list_widget` to `ui/mod.rs` reduce the use of scope-initialization for variables
