+++ spotify_player/src/ui/popup.rs
@@ -1,4 +1,4 @@
-use super::{help, Frame};
+use super::{construct_list_widget, help, Frame};
 use crate::state::*;
 use tui::{layout::*, style::*, widgets::*};
 
@@ -17,6 +17,7 @@ pub fn render_popup(
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Length(7), Constraint::Min(0)].as_ref())
                 .split(rect);
+
             help::render_commands_help_widget(frame, ui, state, chunks[1]);
             (chunks[0], false)
         }
@@ -25,22 +26,20 @@ pub fn render_popup(
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Min(0), Constraint::Length(5)].as_ref())
                 .split(rect);
-            frame.render_stateful_widget(
-                {
-                    let current_device_id = match player.playback {
-                        Some(ref playback) => &playback.device.id,
-                        None => "",
-                    };
-                    let items = player
-                        .devices
-                        .iter()
-                        .map(|d| (format!("{} | {}", d.name, d.id), current_device_id == d.id))
-                        .collect();
-                    construct_list_widget(ui, items, "Devices")
-                },
-                chunks[1],
-                ui.popup.get_list_state_mut().unwrap(),
-            );
+            let widget = {
+                let current_device_id = match player.playback {
+                    Some(ref playback) => &playback.device.id,
+                    None => "",
+                };
+                let items = player
+                    .devices
+                    .iter()
+                    .map(|d| (format!("{} | {}", d.name, d.id), current_device_id == d.id))
+                    .collect();
+                construct_list_widget(ui, items, "Devices", true)
+            };
+
+            frame.render_stateful_widget(widget, chunks[1], ui.popup.get_list_state_mut().unwrap());
             (chunks[0], false)
         }
         PopupState::ThemeList(ref themes, _) => {
@@ -48,14 +47,12 @@ pub fn render_popup(
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Min(0), Constraint::Length(7)].as_ref())
                 .split(rect);
-            frame.render_stateful_widget(
-                {
-                    let items = themes.iter().map(|t| (t.name.clone(), false)).collect();
-                    construct_list_widget(ui, items, "Themes")
-                },
-                chunks[1],
-                ui.popup.get_list_state_mut().unwrap(),
-            );
+            let widget = {
+                let items = themes.iter().map(|t| (t.name.clone(), false)).collect();
+                construct_list_widget(ui, items, "Themes", true)
+            };
+
+            frame.render_stateful_widget(widget, chunks[1], ui.popup.get_list_state_mut().unwrap());
             (chunks[0], false)
         }
         PopupState::UserPlaylistList(_) => {
@@ -63,18 +60,16 @@ pub fn render_popup(
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Min(0), Constraint::Length(10)].as_ref())
                 .split(rect);
-            frame.render_stateful_widget(
-                {
-                    let items = player
-                        .user_playlists
-                        .iter()
-                        .map(|p| (p.name.clone(), false))
-                        .collect();
-                    construct_list_widget(ui, items, "User Playlists")
-                },
-                chunks[1],
-                ui.popup.get_list_state_mut().unwrap(),
-            );
+            let widget = {
+                let items = player
+                    .user_playlists
+                    .iter()
+                    .map(|p| (p.name.clone(), false))
+                    .collect();
+                construct_list_widget(ui, items, "User Playlists", true)
+            };
+
+            frame.render_stateful_widget(widget, chunks[1], ui.popup.get_list_state_mut().unwrap());
             (chunks[0], false)
         }
         PopupState::UserFollowedArtistList(_) => {
@@ -82,18 +77,16 @@ pub fn render_popup(
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Min(0), Constraint::Length(10)].as_ref())
                 .split(rect);
-            frame.render_stateful_widget(
-                {
-                    let items = player
-                        .user_followed_artists
-                        .iter()
-                        .map(|a| (a.name.clone(), false))
-                        .collect();
-                    construct_list_widget(ui, items, "User Followed Artists")
-                },
-                chunks[1],
-                ui.popup.get_list_state_mut().unwrap(),
-            );
+            let widget = {
+                let items = player
+                    .user_followed_artists
+                    .iter()
+                    .map(|a| (a.name.clone(), false))
+                    .collect();
+                construct_list_widget(ui, items, "User Followed Artists", true)
+            };
+
+            frame.render_stateful_widget(widget, chunks[1], ui.popup.get_list_state_mut().unwrap());
             (chunks[0], false)
         }
         PopupState::UserSavedAlbumList(_) => {
@@ -101,18 +94,16 @@ pub fn render_popup(
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Min(0), Constraint::Length(10)].as_ref())
                 .split(rect);
-            frame.render_stateful_widget(
-                {
-                    let items = player
-                        .user_saved_albums
-                        .iter()
-                        .map(|a| (a.name.clone(), false))
-                        .collect();
-                    construct_list_widget(ui, items, "User Saved Albums")
-                },
-                chunks[1],
-                ui.popup.get_list_state_mut().unwrap(),
-            );
+            let widget = {
+                let items = player
+                    .user_saved_albums
+                    .iter()
+                    .map(|a| (a.name.clone(), false))
+                    .collect();
+                construct_list_widget(ui, items, "User Saved Albums", true)
+            };
+
+            frame.render_stateful_widget(widget, chunks[1], ui.popup.get_list_state_mut().unwrap());
             (chunks[0], false)
         }
         PopupState::ArtistList(ref artists, _) => {
@@ -120,10 +111,11 @@ pub fn render_popup(
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Min(0), Constraint::Length(5)].as_ref())
                 .split(rect);
+
             frame.render_stateful_widget(
                 {
                     let items = artists.iter().map(|a| (a.name.clone(), false)).collect();
-                    construct_list_widget(ui, items, "Artists")
+                    construct_list_widget(ui, items, "Artists", true)
                 },
                 chunks[1],
                 ui.popup.get_list_state_mut().unwrap(),
@@ -135,6 +127,7 @@ pub fn render_popup(
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Min(0), Constraint::Length(3)].as_ref())
                 .split(rect);
+
             render_search_box_widget(frame, ui, chunks[1], format!("/{}", query));
             (chunks[0], true)
         }
@@ -148,29 +141,4 @@ fn render_search_box_widget(frame: &mut Frame, ui: &UIStateGuard, rect: Rect, qu
             .title(ui.theme.block_title_with_style("Search")),
     );
     frame.render_widget(search_box, rect);
-}
-
-fn construct_list_widget<'a>(
-    ui: &UIStateGuard,
-    items: Vec<(String, bool)>,
-    title: &str,
-) -> List<'a> {
-    List::new(
-        items
-            .into_iter()
-            .map(|(s, is_active)| {
-                ListItem::new(s).style(if is_active {
-                    ui.theme.current_active()
-                } else {
-                    Style::default()
-                })
-            })
-            .collect::<Vec<_>>(),
-    )
-    .highlight_style(ui.theme.selection_style(true))
-    .block(
-        Block::default()
-            .title(ui.theme.block_title_with_style(title))
-            .borders(Borders::ALL),
-    )
 }
